Somente para Back End

+++ src/main/java/br/com/CalculcadoraCr/model/CargaHoraria.java
@@ -1,20 +1,20 @@
 package br.com.CalculcadoraCr.model;
 
 public class CargaHoraria {
-    private Double horas;
+    private int horas;
 
     public CargaHoraria() {
     }
 
-    public CargaHoraria(Double horas) {
+    public CargaHoraria(int horas) {
         this.horas = horas;
     }
 
-    public double getHoras() {
+    public int getHoras() {
         return horas;
     }
 
-    public void setHoras(Double horas) {
+    public void setHoras(int horas) {
         this.horas = horas;
     }
 }
